Fix uint64 conversion (#7)
* feat: Update Chat model and handler to use uint64 for IDs, and bump version to 0.1.2

Closes #5

* Update internal/handler/chat_handler.go
* Update internal/handler/chat_handler.go
> If the organization ID is already a uint64 in the context, assert it directly to uint64 instead of converting from uint.
* Update internal/handler/chat_handler.go
> Similarly, directly assert orgID as uint64 if possible to simplify the conversion and ensure consistency.
* feat: Update Message and Chat models to use uint64 for IDs, and adjust related service and repository methods
* Refactor codebase for improved readability and maintainability

- Added comments to clarify the purpose of functions and methods across various repositories and services.
- Updated function documentation to follow a consistent style, ensuring clarity on functionality.
- Enhanced code formatting for better readability, including consistent spacing and line breaks.
- Introduced a new configuration file for golangci-lint to enforce coding standards and improve code quality.
- Minor adjustments to error handling and logging for better debugging and traceability.

* refactor: Simplify output format configuration in golangci-lint settings
* refactor: Update golangci-lint configuration to streamline directory exclusions and output formats
* refactor: Update golangci-lint configuration to include issues-exit-code option and enable result sorting
* refactor: Remove timeout and issues-exit-code argument from golangci-lint action
* refactor: Remove .golangci.yml configuration file
* refactor: Change ID type from uint to uint64 across domain, repository, service, and handler layers
* refactor: Simplify OrganizationID assignment in CreateChat and CreateMessage handlers
* Update internal/handler/message_handler.go

The conversion 'uint64(orgID.(uint64))' is redundant since 'orgID' is already asserted as uint64. Consider simplifying the condition to 'if chat.OrganizationID != orgID.(uint64) {'.

---------

+++ internal/domain/apikey.go
@@ -1,35 +1,38 @@
+// Package domain defines the core business entities, interfaces and types
+// that represent the application's domain model for the ChatLogger API.
+// It contains domain entities, service interfaces, and repository interfaces.
 package domain
 
 import (
 	"time"
 )
 
-// APIKey represents authentication credentials for organization API access
+// APIKey represents authentication credentials for organization API access.
 type APIKey struct {
-	ID             uint       `gorm:"primaryKey" json:"id"`
-	OrganizationID uint       `gorm:"not null" json:"organization_id"`
+	ID             uint64     `gorm:"primaryKey"                    json:"id"`
+	OrganizationID uint64     `gorm:"not null"                      json:"organization_id"`
 	HashedKey      string     `gorm:"size:255;uniqueIndex;not null" json:"-"` // Hashed, never return raw
-	Label          string     `gorm:"size:100;not null" json:"label"`
-	CreatedAt      time.Time  `json:"created_at"`
-	RevokedAt      *time.Time `json:"revoked_at,omitempty"`
+	Label          string     `gorm:"size:100;not null"             json:"label"`
+	CreatedAt      time.Time  `                                     json:"created_at"`
+	RevokedAt      *time.Time `                                     json:"revoked_at,omitempty"`
 }
 
-// APIKeyRepository defines the interface for API key data operations
+// APIKeyRepository defines the interface for API key data operations.
 type APIKeyRepository interface {
 	Create(key *APIKey) error
-	FindByID(id uint) (*APIKey, error)
+	FindByID(id uint64) (*APIKey, error)
 	FindByHashedKey(hashedKey string) (*APIKey, error)
-	ListByOrganizationID(orgID uint) ([]APIKey, error)
-	Revoke(id uint) error
-	Delete(id uint) error
+	ListByOrganizationID(orgID uint64) ([]APIKey, error)
+	Revoke(id uint64) error
+	Delete(id uint64) error
 }
 
-// APIKeyService defines the interface for API key business logic
+// APIKeyService defines the interface for API key business logic.
 type APIKeyService interface {
-	GenerateKey(orgID uint, label string) (string, error) // Returns the raw key (only shown once)
+	GenerateKey(orgID uint64, label string) (string, error) // Returns the raw key (only shown once)
 	ValidateKey(rawKey string) (*APIKey, error)
-	GetByID(id uint) (*APIKey, error)
-	ListByOrganizationID(orgID uint) ([]APIKey, error)
-	RevokeKey(id uint) error
-	DeleteKey(id uint) error
+	GetByID(id uint64) (*APIKey, error)
+	ListByOrganizationID(orgID uint64) ([]APIKey, error)
+	RevokeKey(id uint64) error
+	DeleteKey(id uint64) error
 }
